Move tkinter UI related modules to own package
