+++ OTAnalytics/plugin_ui/customtkinter_gui/canvas_observer.py
