+++ OTAnalytics/plugin_ui/customtkinter_gui/constants.py
