+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_analysis.py
@@ -3,7 +3,7 @@ from typing import Any
 from customtkinter import CTkButton, CTkFrame, CTkLabel
 
 from OTAnalytics.adapter_ui.view_model import ViewModel
-from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 
 
 class FrameAnalysis(CTkFrame):
