+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_canvas.py
@@ -10,8 +10,11 @@ from OTAnalytics.adapter_ui.abstract_canvas import AbstractCanvas
 from OTAnalytics.adapter_ui.abstract_frame import AbstractTracksCanvas
 from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.domain.track import TrackImage
-from OTAnalytics.plugin_ui.canvas_observer import CanvasObserver, EventHandler
-from OTAnalytics.plugin_ui.constants import PADX, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.canvas_observer import (
+    CanvasObserver,
+    EventHandler,
+)
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, STICKY
 
 
 @dataclass
