+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_sections.py
@@ -5,7 +5,7 @@ from customtkinter import CTkButton, CTkFrame, CTkLabel
 
 from OTAnalytics.adapter_ui.abstract_treeview import AbstractTreeviewSections
 from OTAnalytics.adapter_ui.view_model import ViewModel
-from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 
 
 class FrameSections(CTkFrame):
