+++ OTAnalytics/plugin_ui/customtkinter_gui/messagebox.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from customtkinter import CTkButton, CTkLabel, CTkToplevel
 
-from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
+from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 
 
 class InfoBox(CTkToplevel):
